pyproject.toml
[project]
name = "pydantic-playgrounds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "mypy>=1.14.1",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic-ai>=0.0.36",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
]
